Move the test exception hierarchy into a common package
To allow them to be reused for circuitbreaker tests
