+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/exception/hierarchy/E2S.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package org.eclipse.microprofile.fault.tolerance.tck.retry.exception.hierarchy;
+package org.eclipse.microprofile.fault.tolerance.tck.exception.hierarchy;
 
 /**
  * Subclass of E2.
