+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/exception/hierarchy/package-info.java
@@ -9,7 +9,7 @@
  * You may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,10 +17,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package org.eclipse.microprofile.fault.tolerance.tck.retry.exception.hierarchy;
 
 /**
- * Subclass of E0 that isn't subclass of E1.
+ * A hierarchy of test exceptions
+ * <p>
+ * Exceptions in this package:
+ * 
+ * <pre>
+ * The <: symbol denotes the subtyping relation (Foo <: Bar means "Foo is a subtype of Bar")
+ * Note that subtyping is reflexive (Foo <: Foo)
+ * 
+ * E0  <: Exception
+ * E1  <: E0
+ * E2  <: E1
+ * E2S <: E2
+ * E1S <: E1, but not E1S <: E2
+ * E0S <: E0, but not E0S <: E1
+ * </pre>
  */
-public class E0S extends E0 {
-}
+package org.eclipse.microprofile.fault.tolerance.tck.exception.hierarchy;
+
